Output spritesheet and index file
This actually makes Spreet useful, as it will now convert the SVGs to PNGs and
pack them into a spritesheet using the rectpack2D bin-packing algorithm
provided by the rectangle-pack library.

This gives Spreet feature-parity with Mapbox's spritezero-cli, although it's not
yet a drop-in replacement — the command-line arguments are provided in a
different order, for example.