Add GitHub Actions workflow to build new version
Run when a tag formatted like a version number is pushed (e.g. v1.0.0).
Spreet is built for five platforms: Linux (x86 64-bit, ARM), MacOS (x86,
ARM), and Windows (x86 64-bit). The release is a manual process for now.

This sort of workflow is quite fiddly and it needed a helping hand from
the equivalent workflows from OxiPNG, Starship, and ripgrep:

<https://github.com/shssoichiro/oxipng/blob/master/.github/workflows/deploy.yml>
<https://github.com/starship/starship/blob/master/.github/workflows/deploy.yml>
<https://github.com/BurntSushi/ripgrep/blob/master/.github/workflows/release.yml>