Increase the size of the target bin if needed
The rectangle-pack library won't resize the target bin if it runs out of space.
But if you give it too large a space then it will do its best to use all that
space. We want the most compact form possible.

The solution I've come up with is to start with a square exactly the size of
the sprites' total pixels and expand it in 0.1 increments each time it runs out
of space.