Move bin-packing code from spreet bin main.rs
This leaves main.rs doing very little other than the CLI stuff, leaving
the business logic in a better position to be tested in isolation. The
bin-packing code now lives in a Spritesheet struct. The
`Spritesheet::new` static method is still a little unwieldy, but I can
refactor that when I get a chance.