Add `--unique` command-line argument (#14)
Match spritezero-cli behaviour by supporting a --unique command-line
argument. When given it only includes unique images in the spritesheet
but maps them to multiple names in the index file.

Includes tests to actually check that the output is correct: that the
`--ratio` flag changes the size of the sprites, and that the `--unique`
flag removes duplicate sprites from the spritesheet and correctly
references them by multiple names in the index. The test run output is
now put into a temp directory rather than the root of the repo.

Also includes a non-consuming builder pattern to generate a
`Spritesheet` object.

Fixes #13.