Use tiny-skia and usvg as re-exported from resvg
The tiny-skia and usvg libraries are only used to pass variables to
resvg. In the master branch of resvg these two libraries are
re-exported, and so we can rely on those instead of trying to keep
Spreet's dependencies in sync. This should solve the issue underlying
PRs #16 and #22, and reduce the noise from Dependabot a bit.