Resolve <image> hrefs relative to the SVG itself
When rendering an SVG, without setting the resource_dir option, any
URLs in the SVG are resolved relative to the current working directory
rather than the SVG itself. The SVG spec expects it to be the directory
that contains the SVG file.

Discovered while writing a test to check that PNG input is supported
according to the method mentioned in #60.