Avoid multiple hash lookups
I am NOT certain if this is correct logically - please take a look:

Currently, sprites are iterated by key, so it goes in sorted order (due to it being a `BTreeMap`).

I changed it to iteration over pixmaps - so now `items` have a different order. It seems to be ok, because the `pack_into_po2` seem to resort them according to the largest first anyway, but I am not as familiar with the algo.

Also, a minor code de-duplication and spelling fix