Hide implementation details with module structure
For a relatively small crate like this, I think it is unnecesary verbose to use nested repeated identifiers like `spreet::error::SpreetError`.  Instead, I propose to move all the crate-defined IDs to the root like `spreet::SpreetError`. The only exception is the `resvg` because its a big lib of itself.