Add support for SDF icons (#77)
Mapbox and MapLibre both support SDF icons, sprites that store a signed
distance field in their alpha channel. They're not well documented
(Mapbox calls them re-colourable images and MapLibre doesn't really
mention them) and they're pretty much abandoned as an idea (see the
conversation in mapbox/mapbox-gl-style-spec#444). But nevertheless both
style specs support them and so Spreet should be able to generate SDF
spritesheets.

Stadia Maps's sdf_glyph_renderer crate supports signed distance field
generation and so we can use that to do the heavy lifting. Although in
theory a single spritesheet can mix-and-match SDF sprites with non-SDF
sprites, in practice spritesheets tend to include one or the other but
not both. And so for now I think it's fine for support to be
all-or-nothing: a spritesheet can have all SDF icons, or no SDF icons.
If there's demand in the future, we can support mixing.