Add CI job to check MSRV is valid
This commit adds a new CI job that checks the minimum supported Rust
version (MSRV) is still valid --- that is, that Spreet still compiles
using that version. The job uses the cargo-hack tool to check the MSRV
against the Rust version specified in the Cargo.toml file.

The job is defined as a GitHub Action and runs on every push.