Add --spacing flag for sprite separation (#95)
Fixes #76. Adds a --spacing argument to apply padding between sprites in
generated spritesheets. This can be used to workaround a problem with
artefacts and bleeding that can occur when tools don't do a good job of
exporting SVG assets.

The spacing is applied to the right and bottom of each sprite during the
packing process, with unused edge spacing trimmed from the final
spritesheet. Sprite dimensions in the JSON index file exclude spacing to
maintain accurate heights and widths.